requirements.txt
ansimarkup==2.1.0
asgiref==3.7.2
aspose-words==23.12.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
click-repl==0.3.0
colorama==0.4.6
compoundfiles==0.3
compressed-rtf==1.0.6
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2.8
django-admin-sortable2==2.2.1
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.7
gron==1.3.0
gunicorn==21.2.0
halo==0.0.31
idna==3.6
inflection==0.5.1
Jinja2==3.1.2
log-symbols==0.0.14
lxml==4.9.3
MarkupSafe==2.1.3
more-itertools==10.1.0
msg-parser==1.2.0
oauthlib==3.2.2
olefile==0.47
packaging==23.2
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-docx==1.1.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
spinners==0.0.24
sqlparse==0.4.4
structlog==23.2.0
sublime-cli==0.0.30
termcolor==2.4.0
typing_extensions==4.8.0
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.12
websockets==12.0
